Move files & add new files
