+++ src/main/scripts/src/app/workspace/organization/permissions/organization-permissions.component.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en">
+   <head>
+      <meta charset="UTF-8">
+      <title>Title</title>
+   </head>
+   <body>
+      
+   </body>
+</html>
